+++ NEWS.md
@@ -1,3 +1,5 @@
+# textrecipes (development version)
+
 # textrecipes 0.2.3
 
 # textrecipes 0.2.2